Create Ezeigwe Blessing Uzoma 2024714157
This python program implements a grading system 'if and 'elif statements. It takes a student's score as input and prints the corresponding grade based on the following scale 

-70-100: Grade A
-60-69: Grade B
-50-59: Grade C
-45-49: Grade D
-40-44: Grade E
-0-39: Grade F

For each valid score,the program also prints 'YES' to confirm the grading was successful. It prints "invalid score"and "NO"

Created by: Ezeigwe Blessing Uzoma 
Reg No: 2024714157